NLP_MLOps&Serving/04_MLOps_and_Serving/02_fastapi-advanced-langgraph/requirements 2.txt
fastapi
uvicorn
pydantic

openai
python-dotenv

chromadb
sentence-transformers
streamlit
requests
